+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### Fixed
 
 - Show line-breaks without "¶" in the Diff Viewer when "Display new-line characters in the Diff Viewer" option is not checked. (#181)
+- Fix window size issue on Windows if the transparency is 100 and the window was maximized in the last session.
 
 ## v6.0